* bootstrap renderer : ** error as result of undo / redo operation will now disappear when html5 history is enabled and back button was pressed ** toolbar flipping (occured in rare cases) fixed when resize event occurs
** crud table : "track by $index" appended to ng repeat over the rows to prevent "duplicates in a repeater are not allowed.". see https://github.com/angular/angular.js/pull/2505 and https://groups.google.com/forum/#!topic/angular/vaoUyP4UgA8

* crud table, crud paginator : both directives will write property "$element" to its models containing a query instance pointing to the real html view

* crud paginator : overall setter/getter "enabled" taking a function as argument or returning the current enabled getter function added. the given enabled function gets a paginator transition as argument returning if this transition is enabled or not when called. this context is the paginator.
example usage :
...
paginator.enabled( function( transition) {
	return !state.list.getEditingContext();
});
...

* state option "ampere.history.html5.hash" added.

function value is expected to return the hash to append to the current document location. "this" context is the state.if function is called with an argument (a hash string) its expected to return either a function proceeding to the given state and data or false/undefined/null for signalling that this state is not responsible to handle this hash.

...
state.option({ 'ampere.history.html5.hash' : function( hash) {
		if( arguments.length) {
				// return function to proceed to state if has matches #person/<id>
			var matches = hash.match( /^person(\/(\d+))?$/);
			return matches && (function( id) {
				return function() {
						// this context is the ampere controller
					return this.proceed( m.transitions.oem_list);
				};
			})( matches[2]);
		} else {
				// return hash
			return "person/" + state.item.id;
		}
	}
});

* state option "ampere.history.html5.title" added

this option is taken into account when option "ampere.history.html5" is enabled. it gives you the option to set a custom document.title used when a browser bookmark is added.
option "ampere.history.html5.title" can be used to override the default default.document title computed using he current state's 'ampere.ui.caption' option.
option value is expected to contain a string or function <string>function() (this context is the state) returning the current document title. if not given the default behaviour (-> current state's 'ampere.ui.caption' option) will be used.

* the ampere controller is now a deferred reflecting the ready state of the ampere widget instance (created with $.ampere( ...)). this context ist the ampere controller.

example use
$( 'body').ampere()
.done( function( module) {
	// do something when ampere widget instance is ready
})

* module option 'ampere.history.html5.deeplinking' added.

deeplinking option (default is enabled). if ampere.history.html5 is true
and the ampere widget element === document.body this function will be called at
startup to resolve a state responsible for handling the hash fragment of
document.location (via state option 'ampere.history.html5').

this option can be overridden with a custom function.
this context is the ampere controller. the given function may return
a deferred tracking its progress.

* module event 'ampere.transition' gets triggered whenever a transition is about to be executed.

the event handler arguments are ( event, transitionDeferred, action, source, target, view). transitionDeferred transports a deferred reflecting the state of the transition execution.
the transitionDeferred handler context is the module. in case the transition succeeds the success handler argument is the previous view. the error handler arguments are the error arguments (xhr, throwable, whatever)

* module event "ampere.transition-enabled" renamed to "ampere.transition-enabled" to prevent jquery event namespace clashes
* module event "ampere.view.changed" renamed to "ampere.view-changed" to prevent jquery event namespace clashes
* module event "ampere.view.updated" renamed to "ampere.view-updated" to prevent jquery event namespace clashes
* module event "ampere.view.refreshed" renamed to "ampere.view-refreshed" to prevent jquery event namespace clashes

* controller.proceed( transition) now returns a deferred tracking the completeness of the state/view switching whenever a transition gets executed.